models/image_caption.py
"""
================================================================================
IMAGE CAPTIONING MODEL - BLIP-BASED VISION-TO-TEXT GENERATION
================================================================================

OVERVIEW:
This module implements an image captioning model using BLIP (Bootstrapping 
Language-Image Pre-training), a state-of-the-art vision-language model. It 
generates natural language descriptions of images by combining computer vision 
and natural language processing capabilities. This demonstrates multimodal AI 
integration within an Object-Oriented Programming framework.

CORE FUNCTIONALITY:
- Automatic image description generation from visual input
- Vision-language understanding using transformer architecture
- Conditional text generation with optional prompt guidance  
- Multi-beam search for improved caption quality
- Support for various image formats and modes

MODEL DETAILS:
- Architecture: BLIP (Bootstrapping Language-Image Pre-training)
- Pre-trained Model: "Salesforce/blip-image-captioning-base"
- Input: PIL Images (any format, automatically preprocessed)
- Output: Natural language text descriptions
- Device Support: CUDA (GPU) with FP16 optimization, CPU fallback

BLIP ARCHITECTURE COMPONENTS:
- Vision Transformer (ViT): Encodes image features
- Text Transformer: Generates captions autoregressively  
- Cross-modal Attention: Aligns visual and textual representations
- Multimodal Fusion: Combines vision and language understanding

OOP CONCEPTS DEMONSTRATED:
1. INHERITANCE:
   - Extends BaseAIModel abstract base class
   - Inherits common model lifecycle management
   - Overrides abstract methods with specialized implementations

2. ENCAPSULATION:
   - Private model components (_processor, _model, _device)
   - Internal state management with controlled access
   - Method hiding for implementation details

3. POLYMORPHISM:
   - Implements BaseAIModel.process() for image captioning
   - Method overloading with different parameter signatures
   - Runtime method dispatch based on input types

4. DECORATOR PATTERN:
   - @validate_input: Ensures proper input format
   - @log_operation: Automatic operation logging
   - @cache_result: Result caching for performance
   - @error_handler: Graceful error management

ADVANCED FEATURES:
- Conditional text generation with user-provided prompts
- Image feature analysis for metadata extraction
- Beam search decoding for higher quality captions
- Mixed precision (FP16) for GPU memory efficiency
- Result caching to avoid redundant computations

PROCESSING METHODS:
1. process(): Standard image-to-caption generation
2. process_with_conditional_text(): Caption with text prompts
3. analyze_image_features(): Extract image metadata and properties

TECHNICAL IMPLEMENTATION:
- Hugging Face Transformers integration for model loading
- Automatic device detection and model placement
- Mixed precision computation for memory efficiency
- Tokenization and preprocessing via BlipProcessor
- Beam search generation with configurable parameters

GENERATION PARAMETERS:
- max_length: Maximum caption length (default: 50 tokens)
- num_beams: Beam search width (default: 5 beams)
- early_stopping: Terminate when end token is generated
- skip_special_tokens: Clean output without special tokens

PERFORMANCE OPTIMIZATIONS:
- Lazy model loading (loaded only when needed)
- GPU acceleration with automatic device selection
- Mixed precision (FP16) to reduce memory usage
- Gradient computation disabled during inference
- Efficient tensor operations with proper device placement

ERROR HANDLING:
- Comprehensive input validation with type checking
- Device compatibility error handling
- Model loading failure recovery
- Memory management for large models

USAGE EXAMPLES:
    # Basic captioning
    model = ImageCaptionModel()
    caption = model.process(pil_image)
    
    # Conditional captioning
    caption = model.process_with_conditional_text(pil_image, "A photo of")
    
    # Image analysis
    features = model.analyze_image_features(pil_image)

INTEGRATION:
- GUI integration through MainWindow class
- Background processing for responsive interface
- File handling integration for image loading/saving
- Logging integration for operation tracking

REFERENCES:
- ChatGPT-5: Implementation assistance and documentation review
- BLIP Model: https://huggingface.co/Salesforce/blip-image-captioning-base
- Hugging Face Transformers: https://huggingface.co/docs/transformers/
- W3Schools Python Classes: https://www.w3schools.com/python/python_classes.asp
- Python Decorators Tutorial: https://realpython.com/primer-on-python-decorators/
================================================================================
"""

# Image captioning model using BLIP

from PIL import Image
import torch
from transformers import BlipProcessor, BlipForConditionalGeneration
from models.base_model import BaseAIModel
from gui.decorators import validate_input, log_operation, cache_result, error_handler


class ImageCaptionModel(BaseAIModel):
    # Image captioning model class
    
    def __init__(self, model_name: str = "Salesforce/blip-image-captioning-base"):
        # Initialize image captioning model
        super().__init__(model_name)
        self._processor = None
        self._model = None
        self._device = "cuda" if torch.cuda.is_available() else "cpu"
    
    def _load_model(self) -> None:
        # Load the BLIP model
        try:
            # Load BLIP processor and model
            self._processor = BlipProcessor.from_pretrained(self._model_name)
            self._model = BlipForConditionalGeneration.from_pretrained(
                self._model_name,
                torch_dtype=torch.float16 if self._device == "cuda" else torch.float32
            )
            self._model = self._model.to(self._device)
            
        except Exception as e:
            raise RuntimeError(f"Failed to load image captioning model: {str(e)}")
    
    @validate_input  # Decorator for input validation
    @log_operation   # Decorator for logging
    @error_handler("Image captioning failed")  # Error handling decorator
    def process(self, image: Image.Image) -> str:
        # Generate caption for image
        self._ensure_loaded()
        
        # Process image and generate caption
        inputs = self._processor(image, return_tensors="pt").to(self._device)
        
        with torch.no_grad():
            generated_ids = self._model.generate(
                **inputs,
                max_length=50,
                num_beams=5,
                early_stopping=True
            )
        
        caption = self._processor.decode(generated_ids[0], skip_special_tokens=True)
        return caption
    
    @cache_result  # Decorator for caching results
    def process_with_conditional_text(self, image: Image.Image, conditional_text: str = "") -> str:
        # Generate caption with conditional text
        self._ensure_loaded()
        
        # Use conditional text if provided
        if conditional_text:
            inputs = self._processor(image, conditional_text, return_tensors="pt").to(self._device)
        else:
            inputs = self._processor(image, return_tensors="pt").to(self._device)
        
        with torch.no_grad():
            generated_ids = self._model.generate(
                **inputs,
                max_length=50,
                num_beams=5
            )
        
        caption = self._processor.decode(generated_ids[0], skip_special_tokens=True)
        return caption
    
    @log_operation
    def analyze_image_features(self, image: Image.Image) -> dict:
        # Extract image features
        # Basic image analysis
        return {
            "size": image.size,
            "mode": image.mode,
            "format": getattr(image, 'format', 'Unknown'),
            "has_transparency": image.mode in ('RGBA', 'LA', 'P')
        }
    
    def get_model_info(self) -> dict:
        # Get model information
        info = super().get_model_info()
        info.update({
            "model_type": "Image Captioning (BLIP)",
            "device": self._device,
            "input_type": "PIL Image",
            "output_type": "Text Caption",
            "description": "Generates descriptive captions for input images using vision-language models"
        })
        return info