requirements.txt
transformers>=4.37.0
torch>=1.12.0
Pillow>=8.0.0
numpy>=1.21.0
tokenizers>=0.12.0